+++ backend/models/subjects.php
@@ -27,6 +27,24 @@ function updateSubject($conn, $id, $subject_name) {
 }
 
 function deleteSubject($conn, $id) {
+
+    //Chequea que el id no este en academic_history en student_id
+    $sql = "SELECT COUNT(*) as total FROM academic_history WHERE subject_id = ?";
+    $stmt = $conn->prepare($sql);
+    $stmt->bind_param("i", $id);
+    $stmt->execute();
+    $result = $stmt->get_result()->fetch_assoc();
+
+    if ($result['total'] > 0) {
+        echo json_encode([
+            "success" => false,
+            "message" => "No se puede eliminar la materia porque tiene materias asociadas."
+        ]);
+        exit();
+    }
+
+    //si es 0 borra.
+
     $sql = "DELETE FROM subjects WHERE id = ?";
     $stmt = $conn->prepare($sql);
     $stmt->bind_param("i", $id);
